Addressed review feedback
- EditInspector - filter out entities with label
- Moved test case into existing file; update cache files with properties & items required for test cae
